Create a new module called app and move start_app.py there with the start.py name
