+++ bootstrap.py
@@ -151,7 +151,7 @@ if not options.hide_console and os.name == 'nt':
     sys.argv.append("--show-console")  # Windows only: show parent console
 
 print("04. Executing spyder.main()")
-from spyderlib import start_app
+from spyderlib.app import start
 
 time_lapse = time.time()-time_start
 print("Bootstrap completed in " +
@@ -159,4 +159,4 @@ print("Bootstrap completed in " +
     # gmtime() converts float into tuple, but loses milliseconds
     ("%.4f" % time_lapse).split('.')[1])
 
-start_app.main()
+start.main()
