+++ continuous_integration/travis/modules_test.sh
@@ -25,9 +25,6 @@ for f in spyderlib/*.py; do
     if [[ $f == spyderlib/tour.py ]]; then
         continue
     fi
-    if [[ $f == spyderlib/start_app.py ]]; then
-        continue
-    fi
     if [[ $f == spyderlib/pyplot.py ]]; then
         continue
     fi
@@ -40,6 +37,9 @@ done
 
 # Depth 2
 for f in spyderlib/*/*.py; do
+    if [[ $f == spyderlib/app/*.py ]]; then
+        continue
+    fi
     if [[ $f == spyderlib/plugins/*.py ]]; then
         continue
     fi
