+++ scripts/spyder
@@ -1,3 +1,3 @@
 #!/usr/bin/env python
-from spyderlib import start_app
-start_app.main()
+from spyderlib.app import start
+start.main()
