+++ spyderlib/app/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © 2009- The Spyder Development team
+# Licensed under the terms of the MIT License
+# (see spyderlib/__init__.py for details)
+
+"""
+spyderlib.app
+=============
+
+Modules related to starting and restarting the Spyder application
+"""
